MODIFY 하는중 DELETEFOODSERVICE DELETEFOODSERVICEIMPL FOOD-DETAILS FOODMAPPER LISTSTORECONTROLLER MODIFYFOODSERVICE MODIFYFOODSERVICEIMPL REVIEWDTO MODIFYREVIEW.JSP

+++ CARESpring/src/main/java/org/care/dto/ReviewDTO.java
@@ -10,7 +10,7 @@ public class ReviewDTO {
 	private int reviewNo;
 	private int userNo;
 	private int storeNo;
-	private double avgScore;
+	private int avgScore;
 	private String reviewContents;
 	private Date reviewDate;
